Bump to 2.0.0-nullsafety.0 && null-safety migration

+++ lib/src/star_clipper.dart
@@ -17,7 +17,7 @@ class StarClipper extends CustomClipper<Path> {
 
     double radius = halfWidth / 2;
 
-    double degreesPerStep = _degToRad(360 / numberOfPoints);
+    double degreesPerStep = _degToRad(360 / numberOfPoints) as double;
 
     double halfDegreesPerStep = degreesPerStep / 2;
